# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Democra — API interna (server/)
  version: "1.0.0"
  description: >
    Documentación base de la API Express de Democra (server/index.js).
    Extraída de los routers reales (server/routes/*.js), no inventada.
    Todos los endpoints requieren un Bearer token de Supabase Auth salvo
    donde se indique explícitamente lo contrario (terminal-login, validate-ruc).
    Este archivo es el punto de partida ("estructura base") para mantener la
    documentación de la API sincronizada con el código — se debe actualizar
    cada vez que se agregue, cambie o elimine un endpoint en server/routes/.
servers:
  - url: http://localhost:8787/api
    description: Desarrollo local (server/index.js, puerto por defecto de config.js)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Supabase JWT
      description: >
        Access token de una sesión de Supabase Auth. El backend lo resuelve
        vía resolveAuthContext() (server/supabase.js) para obtener user + profile.
  schemas:
    ErrorResponse:
      type: object
      description: Formato estándar de error (server/utils/http.js sendError()).
      properties:
        error_code:
          type: string
          example: "IAM-004"
        error_type:
          type: string
          example: "auth"
        message:
          type: string
...
    Sede:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
  responses:
    Unauthorized:
      description: Token ausente, inválido o expirado.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Body o parámetros inválidos.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
security:
  - bearerAuth: []
paths:
...
  /auth/step-up/verify-otp:
    post:
      tags: [auth]
      summary: Verifica un código OTP de un desafío de step-up ya iniciado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [challenge_id, code]
              properties:
                tenant_id:
                  type: string
                  format: uuid
                challenge_id:
                  type: string
                code:
                  type: string
                  description: Código OTP de 6 dígitos.
      responses:
        "200":
          description: Sesión step-up creada tras verificar el OTP.
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/step-up/resend-otp:
    post:
      tags: [auth]
      summary: Reenvía un código OTP para un desafío de step-up activo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tenant_id:
                  type: string
                  format: uuid
                challenge_id:
                  type: string
      responses:
        "200":
          description: OTP reenviado.
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/terminal-login:
    post:
      tags: [auth]
      summary: Login por PIN en un terminal físico (kiosco), sin sesión Supabase previa
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenant_id, user_id, pin, terminal_id]
              properties:
                tenant_id:
                  type: string
                  format: uuid
                user_id:
                  type: string
                  format: uuid
                pin:
                  type: string
                terminal_id:
                  type: string
                  format: uuid
                user_agent:
                  type: string
      responses:
        "200":
          description: Sesión de terminal creada.
        "400":
          $ref: "#/components/responses/ValidationError"
        "403":
          description: Terminal inactivo o no encontrado (IAM-005).
...
  /onboarding/validate-ruc/{ruc}:
    get:
      tags: [onboarding]
      summary: Valida un RUC de 11 dígitos contra el servicio fiscal externo configurado
      security: []
      parameters:
        - name: ruc
          in: path
          required: true
          schema:
            type: string
            pattern: "^\\d{11}$"
      responses:
        "200":
          description: Datos fiscales del RUC.
        "400":
          description: RUC con formato inválido (TEN-001).
        "503":
          description: Servicio de validación fiscal no configurado.
  /onboarding/bootstrap-tenant:
    post:
      tags: [onboarding]
      summary: Crea (o devuelve, si ya existe) el tenant del usuario autenticado
      description: >
        Envuelve public.fn_bootstrap_tenant: crea tenant + profile + sede
        "Principal" + rol Owner + asignación, de forma idempotente. Antes de
        este endpoint no había ningún camino (Express ni Supabase directo)
        para que una organización nueva se diera de alta.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenant_name, tax_id, industry_type_id]
              properties:
                tenant_name:
                  type: string
                tax_id:
                  type: string
                  pattern: "^\\d{11}$"
                  description: RUC de 11 dígitos.
                industry_type_id:
                  type: string
                  example: ong
                plan_id:
                  type: string
                  default: basic
                billing_day:
                  type: integer
                  minimum: 1
                  maximum: 28
                  default: 1
      responses:
        "201":
          description: Tenant creado o ya existente para este usuario.
          content:
            application/json:
              schema:
                type: object
                properties:
                  tenant_id:
                    type: string
                    format: uuid
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /sedes:
    get:
      tags: [sedes]
      summary: Lista las sedes del tenant actual
      responses:
        "200":
          description: Lista de sedes.
          content:
            application/json:
              schema:
                type: object
                properties:
                  sedes:
                    type: array
                    items:
                      $ref: "#/components/schemas/Sede"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [sedes]
      summary: Crea una nueva sede (requiere ser administrador del tenant)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        "201":
          description: Sede creada.
          content:
            application/json:
              schema:
                type: object
                properties:
                  sede:
                    $ref: "#/components/schemas/Sede"
        "400":
          $ref: "#/components/responses/ValidationError"
        "403":
          description: El actor no es administrador del tenant (IAM-003).
...